refactor: 스타일 관련 코드 분리 및 디렉터리 생성
